requirements.txt
pandas>=1.3.5
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.2
fastapi>=0.95.0
uvicorn>=0.15.0
psycopg2-binary>=2.9.3
sqlalchemy>=1.4.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=0.20.0
pytest>=7.0.0
pydantic>=1.10.0
scikit-learn>=1.0.0
langchain>=0.0.267
langchain-community>=0.0.1